doc/src/docker-run/requirements.txt
schedule==1.1.0
PyGithub==1.55
